+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - entrypoint `all-extractors`: run all extractors
 
 - transform grippeweb resources
+- wikidata aux extractor into seq-repo
 
 ### Changes
 
@@ -37,7 +38,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - implement odk transform functions
 - grippeweb extract
 - dagster schedules for non-default groups, configurable via setting `schedule`
-- wikidata aux extractor into seq-repo
 
 ### Changes