+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
-- update mex-common to 0.26.0
+- update mex-common to 0.26.1
 - move `mex.pipeline` documentation to `__init__` to have it in sphinx
 
 ### Deprecated